Move time picker state to composable
